pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "starfish-mcp"
version = "0.1.0"
description = "MCP server for Starfish Storage integration"
authors = [
    {name = "Starfish Storage", email = "info@starfishstorage.com"}
]
dependencies = [
    "mcp>=1.0.0",
    "aiohttp>=3.9.0",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "pydantic>=2.0.0",
    "structlog>=24.1.0"
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "Apache-2.0"}

[project.scripts]
starfish-mcp = "starfish_mcp.server:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0"
]

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.black]
line-length = 88
target-version = ["py38"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
